Rename app_test to e2e_test
